support local predictor for quantized model (#2042)
