+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/optim/LocalPredictorSpec.scala
@@ -116,4 +116,24 @@ class LocalPredictorSpec extends FlatSpec with Matchers with BeforeAndAfter {
       assert(imageFeatures(x - 1).predict() != null)
     })
   }
+
+  "predictImage with quantize" should "work properly" in {
+    import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric.NumericFloat
+    RNG.setSeed(100)
+    val resource = getClass.getClassLoader.getResource("pascal/")
+    val imageFrame = ImageFrame.read(resource.getFile) ->
+      Resize(256, 256) -> CenterCrop(224, 224) ->
+      ChannelNormalize(0.485f, 0.456f, 0.406f, 0.229f, 0.224f, 0.225f) ->
+      MatToTensor() -> ImageFrameToSample()
+    val model = Inception_v1_NoAuxClassifier(classNum = 20).quantize()
+    val detection = model.predictImage(imageFrame).toLocal()
+    val feature = detection.array.head
+    println(feature(ImageFeature.predict))
+
+    val imageFeatures = detection.array
+    val prob = imageFeatures.map(x => x[Tensor[Float]](ImageFeature.predict))
+    val data = imageFeatures.map(_[Sample[Float]](ImageFeature.sample))
+    prob(0) should be(model.evaluate().forward(data(0).feature.reshape(Array(1, 3, 224, 224)))
+      .toTensor[Float].split(1)(0))
+  }
 }
